ft_put* upgrade (int <- ssize_t)

+++ 112.114.105.110.116.102/srcs/ft_putendl.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int		ft_putendl(char const *s)
+ssize_t		ft_putendl(char const *s)
 {
 	return (ft_putendl_fd(s, 1));
 }
