+++ 112.114.105.110.116.102/srcs/ft_puterr.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int		ft_puterr(char const *s)
+ssize_t		ft_puterr(char const *s)
 {
 	return (ft_putstr_fd(s, 2));
 }
